requirements.txt
streamlit
pandas
unstract-llmwhisperer